requirements.txt
ttkbootstrap==1.10.1
playsound==1.2.2
pillow==9.4.0
